[heatpumpir] Fix ambiguous millis() call with HeatpumpIR library
The HeatpumpIR library's HeatpumpIRCompat.h header defines millis() in
the global namespace, which conflicts with esphome::millis() when both
are visible. This caused compilation failures on ESP32-IDF builds.

Fix by moving the IRSenderESPHome class directly into heatpumpir.cpp,
keeping the HeatpumpIR library headers isolated to a single translation
unit instead of leaking through esphome.h.
